[REFACTOR] Rename dashboard pages to views
