+++ src/client/react/user/RegistrationClosed/Dashboard.js
@@ -4,10 +4,10 @@ import { connect } from 'react-redux';
 import { Switch, Redirect } from 'react-router-dom';
 import Base from '../components/Base';
 import HelpMenu from '../components/HelpMenu';
-import Home from '../pages/Home';
-import Help from '../pages/Help';
-import Instructions from '../pages/Instructions';
-import NotFound from '../pages/NotFound';
+import Home from '../views/Home';
+import Help from '../views/Help';
+import Instructions from '../views/Instructions';
+import NotFound from '../views/NotFound';
 
 
 const mapStateToProps = (state, ownProps) => {
