+++ src/client/react/user/views/Home/TeamPanel.js
