+++ src/client/react/user/views/Instructions/InstructionCollapse.js
