+++ src/client/react/user/views/Instructions/index.js
